w6/advanced/requirements.txt
streamlit
langchain
openai
python-dotenv
PyGithub
langchain-community
langchain-openai